<bugfix> Fix scale bug
This commit fixes a bug where scaling an object sometimes causes a red border to appear along the edges. This bug occurs due to the dot product producing negative values due to floating point error when the incident ray is nearly orthogonal to the surface normal. The bug was solved by clamping these calculations.